requirements.txt
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
ldap3==2.8
nose==1.3.7
nose-parameterized==0.6.0
psycopg2-binary==2.8.5